plat/include: Define address offsets of boot stack and pagetable
If we place the boot stack and pagetable in BSS section. These
areas are not easy to be reused after changing to new stack.
So, in Arm64, we want to place the pagetable and boot stack
after the end of image. In this case, once we change to new
stack or we have new pagetable, these two areas can be reclaimed
very easy.

Note that, currently, this memory layout applys to QEMU/KVM only.

Signed-off-by: Wei Chen <Wei.Chen@arm.com>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>