plat/common: Add cache maintenance support for arm64
When MMU is disabled, when we modify some data, for example, write
page-table, we will write with Device nGnRnE attributes. So the
cache will be bypassed. But the cache may still contain stall data
that we will hit when enabling MMU and cache. To prevent such issue,
we need to clean the cache potentially before and after updating
the page-table area.

So we introduce the cache maintenance functions in this patch.

Signed-off-by: Wei Chen <wei.chen@arm.com>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>