plat/common: Add counter workaround for Cortex-A73 erratum 858921
The errata #858921 describes that Cortex-A73 (r0p0 - r0p2)
counter read can return a wrong value when the counter crosses
a 32bit boundary, but newer Cortex-A73 are not affected.

The workaround involves performing the read twice, compare
bit[32] of the two read values. If bit[32] is different,
keep the first value, otherwise keep the second value.

Signed-off-by: Wei Chen <wei.chen@arm.com>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>