plat/kvm: Parse command line from device tree for Arm64
Device tree has a "chosen" node, this node can be used to pass
parameters to virtual machine through bootargs property. We can
use this feature to pass command line to unikraft application,
and then parse it from device tree for further usage.

The command line will be modified in uk_argparse, but the memory
attribute of DTB is readonly, we need a copy of this command line.

Signed-off-by: Wei Chen <Wei.Chen@arm.com>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>