plat/xen: add PVH console initialization
Contrary to PV, console page isn't mapped on boot. Since we want console as
early as possible, reserve a page for it in initial memory layout and then
remap the page we get from Xen over it. This way, it can be done before memory
allocator is set. Make get_pte() work very early to make it possible.